+++ packages/storybook/examples/form/SelectList/MultipleUncontrolled.js
@@ -8,7 +8,7 @@ function MultipleUncontrolled() {
     return (
         <SelectList
             multiple
-            defaultValues={['1']}
+            defaultValue={['1']}
             onChange={action('change')}>
             <Option label="Option A" value="1" />
             <Option label="Option B" value="2" />